implemented Archives model, updated archives routes, updated and fixed user aside bar with archives menu

+++ resources/views/layouts/diplomas.blade.php
@@ -20,14 +20,16 @@
 					    		<td class="col-md-10">{{ $diploma->diploma_type }}
 					    			<p><a class="btn" href="{{  action('DiplomaController@show', [$diploma->id]) }}">view details »</a></p>
 					    		</td>
-					    		<td class="col-md-1"><a href="{{ action('DiplomaController@edit', [$diploma->id]) }}" class="glyphicon glyphicon-pencil default"></a></td>
-					    		<td>
-						    		{{ Form::model( $diploma, ['method'=>'DELETE', 'action' => ['DiplomaController@destroy', $diploma->id]] ) }}
-					                    {{ Form::hidden('_method', 'DELETE') }}
-					                    <button type="submit"><i class="glyphicon glyphicon-trash pull-right"></i></button>
-					                    <!-- {{ Form::submit('delete', array('class' => 'btn btn-danger btn-xs glyphicon glyphicon-trash')) }} -->
-					                {{ Form::close() }}
-					    		</td>
+					    		@if((Auth::check()) AND (Auth::User()->id === $user->id))
+						    		<td class="col-md-1"><a href="{{ action('DiplomaController@edit', [$diploma->id]) }}" class="glyphicon glyphicon-pencil default"></a></td>
+						    		<td>
+							    		{{ Form::model( $diploma, ['method'=>'DELETE', 'action' => ['DiplomaController@destroy', $diploma->id]] ) }}
+						                    {{ Form::hidden('_method', 'DELETE') }}
+						                    <button type="submit"><i class="glyphicon glyphicon-trash pull-right"></i></button>
+						                    <!-- {{ Form::submit('delete', array('class' => 'btn btn-danger btn-xs glyphicon glyphicon-trash')) }} -->
+						                {{ Form::close() }}
+						    		</td>
+						    	@endif
 				    		</div>
 				    		
 				    	</tr>
