+++ resources/views/layouts/newTopic.blade.php
@@ -20,14 +20,17 @@
 					    		<td class="col-md-10">{{ $topic->title }}
 					    			<p><a class="btn" href="{{  action('TopicsController@show', [$topic->id]) }}">view details »</a></p>
 					    		</td>
-					    		<td class="col-md-1"><a href="{{ action('TopicsController@edit', [$topic->id]) }}" class="glyphicon glyphicon-pencil default"></a></td>
-					    		<td>
-						    		{{ Form::model($topic,['method'=>'DELETE', 'action' => ['TopicsController@destroy', $topic->id]]) }}
-					                    {{ Form::hidden('_method', 'DELETE') }}
-					                    <button type="submit"><i class="glyphicon glyphicon-trash pull-right"></i></button>
-					                   
-					                {{ Form::close() }}
-					    		</td>
+
+					    		@if((Auth::check()) AND (Auth::User()->id === $user->id))
+						    		<td class="col-md-1"><a href="{{ action('TopicsController@edit', [$topic->id]) }}" class="glyphicon glyphicon-pencil default"></a></td>
+						    		<td>
+							    		{{ Form::model($topic,['method'=>'DELETE', 'action' => ['TopicsController@destroy', $topic->id]]) }}
+						                    {{ Form::hidden('_method', 'DELETE') }}
+						                    <button type="submit"><i class="glyphicon glyphicon-trash pull-right"></i></button>
+						                   
+						                {{ Form::close() }}
+						    		</td>
+					    		@endif
 				    		</div>
 				    		
 				    	</tr>
