+++ resources/views/layouts/newTraining.blade.php
@@ -22,14 +22,17 @@
 					    		<td class="col-md-10">{{ $training->topic }}
 					    			<p><a class="btn" href="{{  action('TrainingController@show', [$training->id]) }}">view details »</a></p>
 					    		</td>
-					    		<td class="col-md-1"><a href="{{ action('TrainingController@edit', [$training->id]) }}" class="glyphicon glyphicon-pencil default"></a></td>
-					    		<td>
-						    		{{ Form::model($training,['method'=>'DELETE', 'action' => ['TrainingController@destroy', $training->id]]) }}
-					                    {{ Form::hidden('_method', 'DELETE') }}
-					                    <button type="submit"><i class="glyphicon glyphicon-trash pull-right"></i></button>
-					                   
-					                {{ Form::close() }}
-					    		</td>
+					    		
+					    		@if((Auth::check()) AND (Auth::User()->id === $user->id))
+						    		<td class="col-md-1"><a href="{{ action('TrainingController@edit', [$training->id]) }}" class="glyphicon glyphicon-pencil default"></a></td>
+						    		<td>
+							    		{{ Form::model($training,['method'=>'DELETE', 'action' => ['TrainingController@destroy', $training->id]]) }}
+						                    {{ Form::hidden('_method', 'DELETE') }}
+						                    <button type="submit"><i class="glyphicon glyphicon-trash pull-right"></i></button>
+						                   
+						                {{ Form::close() }}
+						    		</td>
+					    		@endif
 				    		</div>
 				    		
 				    	</tr>
